Help reduce build problems due to font declarations. https://bugs.webkit.org/show_bug.cgi?id=36190
Reviewed by Adam Roben.

Functions unchanged, no new tests.

* WebCore.vcproj/WebCore.vcproj: Add new file, and show
  movement of FontPlatformData.h to cg/ directory.
* platform/graphics/cairo/FontPlatformData.h: Merge in
  WinCairo-specific declarations.  Correct constructor
  declaration to use type float for font size.
(WebCore::FontPlatformData::FontPlatformData):
(WebCore::FontPlatformData::hfont):
(WebCore::FontPlatformData::useGDI):
(WebCore::FontPlatformData::fontFace):
(WebCore::FontPlatformData::setSize):
(WebCore::FontPlatformData::hash):
(WebCore::FontPlatformData::isHashTableDeletedValue):
* platform/graphics/cairo/FontPlatformDataCairo.cpp:
(WebCore::FontPlatformData::FontPlatformData): Correct
  constructor to accept type float for font size.
* platform/graphics/cg/FontPlatformData.h: Copied from WebCore/platform/graphics/win/FontPlatformData.h.
(WebCore::FontPlatformData::FontPlatformData):
(WebCore::FontPlatformData::hfont):
(WebCore::FontPlatformData::cgFont):
(WebCore::FontPlatformData::operator==):
* platform/graphics/win/FontPlatformData.h: Removed.
* platform/graphics/win/FontPlatformDataCairoWin.cpp:
(WebCore::FontPlatformData::operator==): Move implementation
  here (previously in header) to match other cairo-based ports.
* platform/graphics/win/RefCountedHFONT.h: Added.  This
  was extracted from win/FontPlatformData.h, and is now shared
  by the cg/FontPlatformData.h and cairo/FontPlatformData.h
(WebCore::RefCountedHFONT::create):
(WebCore::RefCountedHFONT::createDeleted):
(WebCore::RefCountedHFONT::~RefCountedHFONT):
(WebCore::RefCountedHFONT::hfont):
(WebCore::RefCountedHFONT::hash):
(WebCore::RefCountedHFONT::RefCountedHFONT):



Canonical link: https://commits.webkit.org/47486@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@56192 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/platform/graphics/cairo/FontPlatformDataCairo.cpp
@@ -134,7 +134,7 @@ FontPlatformData::FontPlatformData(float size, bool bold, bool italic)
 {
 }
 
-FontPlatformData::FontPlatformData(cairo_font_face_t* fontFace, int size, bool bold, bool italic)
+FontPlatformData::FontPlatformData(cairo_font_face_t* fontFace, float size, bool bold, bool italic)
     : m_pattern(0)
     , m_fallbacks(0)
     , m_size(size)
